Former-commit-id: 201d8237948f76945430817ad9406b26d7f49bfa

+++ applications/plugins/Compliant/mapping/ContactMapping.h
@@ -63,7 +63,7 @@ protected:
 		// local frames have been computed in assemble
 
         (void)in;
-        assert( in.size() == out.size() || std::count( mask.begin(),mask.end(),true)==out.size() );
+        assert( in.size() == out.size() || (size_t)std::count( mask.begin(),mask.end(),true)==out.size() );
 
         unsigned n = out.size();
